+++ less/tests/django_settings.py
@@ -34,3 +34,5 @@ LOGGING = {
         },
     }
 }
+
+SECRET_KEY = "secret"